refactor(glossary): move to own TS project
