+++ glossary/ExpandableInlineBlock.scss
