+++ glossary/tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "extends": "../devTools/tsconfigs/tsconfig.base.json",
+    "compilerOptions": {
+        "outDir": "../itsJustJavascript/glossary",
+        "rootDir": "."
+    },
+    "references": [{ "path": "../clientUtils" }]
+}
