+++ site/formatting.tsx
@@ -12,8 +12,8 @@ import { BAKED_BASE_URL, WORDPRESS_URL } from "../settings/serverSettings.js"
 import { bakeGlobalEntitySelector } from "./bakeGlobalEntitySelector.js"
 import { PROMINENT_LINK_CLASSNAME } from "./blocks/ProminentLink.js"
 import { Byline } from "./Byline.js"
-import { formatGlossaryTerms } from "./formatGlossary.js"
-import { getMutableGlossary, glossary } from "./glossary.js"
+import { formatGlossaryTerms } from "../glossary/formatGlossary.js"
+import { getMutableGlossary, glossary } from "../glossary/glossary.js"
 import { SectionHeading } from "./SectionHeading.js"
 
 export const GRAPHER_PREVIEW_CLASS = "grapherPreview"
