+++ site/owid.scss
@@ -56,12 +56,13 @@
 @import "site/blocks/Grid.scss";
 @import "site/blocks/Card.scss";
 @import "site/blocks/BiographyCard.scss";
-@import "site/ExpandableInlineBlock.scss";
-@import "site/GlossaryExcerpt.scss";
 @import "site/PostCard/PostCard.scss";
 @import "site/Breadcrumb/Breadcrumb.scss";
 @import "site/DataValue.scss";
 @import "site/AnnotatingDataValue.scss";
+
+@import "glossary/GlossaryExcerpt.scss";
+@import "glossary/ExpandableInlineBlock.scss";
 
 /* HACK (Mispy): Fix search autozooming on iPhone Safari.
    http://stackoverflow.com/questions/2989263/disable-auto-zoom-in-input-text-tag-safari-on-iphone */
