+++ tsconfig.json
@@ -10,6 +10,7 @@
         { "path": "./clientUtils" },
         { "path": "./gridLang" },
         { "path": "./gitCms" },
+        { "path": "./glossary" },
         { "path": "./explorer" },
         { "path": "./coreTable" },
         { "path": "./site" },
